Fix error return code for open() and other syscalls on x86-64
On error, the open() syscall was returning errno & 0xffffffff in %rax
on x86-64 instead of -errno.  This stops glibc from regarding this as
an error, and so its syscall wrapper returns -errno instead of -1 and
does not set errno.

I have fixed up the other syscalls to use long (64-bit) instead of int
(32-bit) as well.  Not all of them were affected by the problem: it
depends on gcc's code generation.  Sometimes casting to int truncates
a value, sometimes it doesn't.  It seems better to be consistent
though.

Adds a test for open() and some other syscalls.

Review URL: http://codereview.chromium.org/1648009

git-svn-id: https://seccompsandbox.googlecode.com/svn/trunk@72 55e79e8e-603c-11de-8c10-5fe6993ea61f

+++ clone.cc
@@ -7,8 +7,8 @@
 
 namespace playground {
 
-int Sandbox::sandbox_clone(int flags, void* stack, int* pid, int* ctid,
-                           void* tls, void *wrapper_sp) {
+long Sandbox::sandbox_clone(int flags, void* stack, int* pid, int* ctid,
+                            void* tls, void *wrapper_sp) {
   long long tm;
   Debug::syscall(&tm, __NR_clone, "Executing handler");
   struct {
@@ -51,7 +51,7 @@ int Sandbox::sandbox_clone(int flags, void* stack, int* pid, int* ctid,
     die("Failed to forward clone() request [sandbox]");
   }
   Debug::elapsed(tm, __NR_clone);
-  return static_cast<int>(rc);
+  return rc;
 }
 
 bool Sandbox::process_clone(int parentMapsFd, int sandboxFd, int threadFdPub,
